first experiments with ELO
